requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
motor>=3.3.1
apscheduler>=3.10.4
